+++ sdk/keyvault/Azure.Security.KeyVault.Secrets/CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Breaking Changes
 
 - Added the `outContentType` query parameter to the `SecretClient.GetSecret` and `SecretClient.GetSecretAsync` to specify the format in which the certificate will be returned.
+- Added the `previousVersion` property to `SecretProperties`.
 
 ### Bugs Fixed